Image class: several fixes rolled into one: rigorous error checking for /all/ GD operations to maximize the chance to detect the Swedish Chef ('b0rk'). All errors throw Exceptions to simplify the interface; for the same reason all methods return a reference to the current instance ($this) so they can be chained at will. And the 'memory usage guestimator' has now been integrated: the constructor will check whether there's enough and when not, throw a suitable Exception.
NOTE: anticipating the next several commits and its close relation to the fileManager backend, the Image class ASSUMES the file specified is 'cleaned up', i.e. the Image class constructor MUST be presented with a normalized and otherwise cleaned up PATH to the image file.
This does NOT mean the Image class expects a correct and accessible file! When a non-existing and/or corrupt image file is detected, an appropriate Exception will be thrown. The same goes for very large images which will not fit in RAM, given the current settings; formerly such images would make the PHP engine FAIL SILENTLY (due to fatal out of memory conditions occurring).

The current state of affairs is that silent fatalities due to out of memory conditions happening in the GD lib are gone WITH HIGH PROBABILTY -- heck, I'm sure someone with enough zeal and expertise can craft something that'll thwart the guestimator.

The current implementation strives to cajole the PHP engine into garbage collecting image/GD buffers ASAP to keep the memory footprint as low as possible.