FileManager backend: the big mother: strict authentication checking / user callbacks for all 'non-intrusive' FileManager actions (upload, download, create, move/copy, delete) + now using the new Image class features (mimemap passing, memory guestimator-based checking) + tinyMCE demo files updated accordingly (TODO: CKeditor?)
Now each event calls a user-configurable callback function, which can decide whether that action is permitted or not. The callback is provided with as much info as possible about the event/action and the files / paths involved to allow for complex custom decision logic (when needed).

Now all events/action handlers catch exceptions thrown internally and produce JSON encoded error reports (TODO: client side does not really handle 'em all yet).

Note that the VIEW and DETAIL events have been augmented in this regard:

1) when VIEW encounters an error (most probably due to a directory disappearing due to on-site admin activity or other outside interventions), the FileManager backend will fall back to the configured base directory and try to stimulate the client side JS code to take it from there (this is related to issue #24: even under error coditions, the backend now attempts to provide a reasonable 'root' JSON element, among other things, in order to make the FileManager JS code operate smoothly under nonregular conditions)

2) when DETAIL encounters an error, this is most probably due to a missing, corrupted or overly large image file, and the error feedback (JSON encoded) will try to report such failures intelligently. (TODO: remove the debug output which reports PHP engine memory consumption while looking at thumbnails)

3) in anticipation of the next commit, several path components are urlencoded to allow a wider range of permissible filename and directory characters. (TODO: see what happens when we place the FM on top of a pre-rigged Chinese-locale directory layout: which bit of JS/FM is going to b0rk then?)

The FM backend now has the added options' elements to configure which events/actions are allowed (that is without having to go to the relevant authorization callback!): 'upload' and 'destroy' already existed, 'create' , 'move' and 'download' have been added to the set, mirroring the FM JS side. Note the comment regarding the 'download' option; despite this consideration, it is still useful to have it.

Notes: getRealPath() now by default does NOT create the directory if it does not exists: this is more sensible behaviour from a security point of view. The only spot where 'implied directory creation' is used is with the thumbnails directory: when the constructor code finds it does not exist yet, it will be created.

Also note that this patch should also take care of issue #21, but since that one is tough to reproduce (haven't succeeded here either), cave canem. Based on review, the new code should be even stricter about it then it was before.