be more allowing in what may pass through to the filesystem as characters in a path / file name: also allow various non-alphanumerics in the filename: '., \[\]\(\)~&!@#_+-' is the allowed extra set, but these are not allowed (and hence stripped off) at the start or end of the filename: '_.,&+'
Note that the somewhat cumbersome characters in there are # and & as those required the rawurlencode-based processing of file / path specs before having them JSON encoded: these results are used as part of URIs, after all.

Currently the frontend is ever so slightly intolerant of # and & in file and/or path names: when the path of the currently viewed directory of file contains either of these, a [F5]-driven refresh of the page in the browser will set you back to the basedir as the FM client code does not encode these when storing them in the 'bookmark' part of the URL (which is used by the FM client side as a local store for path and file location tracking).