FM backend: two stability tweaks:
(1) add @ for various filesystem operations which are called while producing the view (there are scenarios where this is required, e.g. when the frontend fires multiple different requests shortly after one another, and one of these is deleting a directory or subtree while another request to produce a view thereof is still pending)

(2) the 'view' event now has a prettier fallback: when a directory view fails, we try the PARENT instead of BASEDIR. We keep on trying to view the parent (of the parent, etc.) until we reach the BASEDIR, until we succeed. If we still fail by then, we stop trying and send a error report.

Note that any level of 'faaling back' in (2) means that the status=0 and the first error discovered will be reported alongside. The frontend does not really watch for these errors, so currently this detail is sorta invisible, apart from the 'odd' dropping back of the view towards the basedir.