backend: fixed a bugs in the Image class, moved the memory usage setting to a clearly visible define() at the top of the file, adjusted the way Image copes with memory usage, changed the resize() default parameter values to more sensible values and added the store_original_if_unaltered parameter to the process() and save() methods so we now can keep animated GIFs intact, for one.
The resize() method had a quirk when ratio==false: in that particular case, scaling was NOT always ensuring the resized result would indeed fit the x/y limits, when one of those limits would be NULL, i.e. an /implied/ ratio==true processing.

Now Image.class.php() will adjust the PHP memory_limit setting only iff both the currently active setting is smaller than the define()d limit AND the guestimator predicts a memory usage which would fit within the bounds of the define()d memory space value: after all, bumping up mem usage limits is only useful when we have a good guess that we'll actually be able to use the increased amount of memory and do some good.
A bug in the old logic was also fixed: previously, Image would bluntly adjust the memory limit but NOT benefit from it as the guestimator signal (will_fit) could not be updated to the new memory availability status.