chopped the 'fill()' routine into time-delayed chunks to prevent the dreaded 'slow script, abort or continue?' dialog in browsers when viewing large directories. The time taken to perform the actions is monitored and the chunks are dimensioned accordingly to allow for the wide range of client (browser machine) performance.
The delay per chunk is absolutely minimal and only intended to relinquish control back to the browser itself to keep it quiet about 'long running scripts' ;-)

When the user clicks a directory or does something else to trigger a new 'view' run, while the old one still hasn't finished, the old one is aborted ASAP (the timer handle is used to $clear the timer before starting with the new fill).

TODO: the current implementation can still show a few items of the 'old' view run in the new, i.e. files can show up which are not really there: this is due to the abort of the old run being executed, but there is no guearantee the very last running 'old' chunk still doesn't get a few items done before quitting.
We'll keep this one as a 'known bug'. (Heck, before FM would simply croak the browser on such large directories, so it's still an improvement if you ask me ;-) )