a very odd JavaScript scope and binding bug, and an apparent fix, though I still don't fully understand why this does work:
WARNING: for some (to me) incomprehensible reason the old code which bound the event handlers to 'this==self' and which used the 'el' variable available here, does NOT WORK ANY MORE - tested in FF3.6. Turns out 'el' is pointing anywhere but where you want it by the time the event handler is executed.

The 'solution' which I found was to rely on the 'self' reference instead and bind to 'el'. If the one wouldn't work, the other shouldn't, but there you have it: this way around it works. FF3.6.14 :-(

TODO: test this 'solution' on other browsers and see how they fare...