frontend: the Final Solution for handling huge directories: pagination:
the pagination is added to limit the maximum number of rows rendered in the list/thumb view as this is a very slow process once you reach certain numbers (1K..2K+ for modern machines serving as browser h/w), which deteriorates user experience.
The new code limits the number of items rendered and, when necessary, shows a few next/previous pagination controls in the top bar above the list. Tooltips are used to describe the purpose of the new controls, while including information about the currently viewed page and the total number of pages.

Last page position is persisted together with filename and path (jsGet).

The code has been designed to insure correct operation when users perform actions while a previous list render is pending (the chunked fill): the previous fill render is always aborted before the next one starts.