several fixes to make TinyMCE work with FM again; now tinyMCE also has the ability to 'jump' to the previously selected image file - a feature introduced with the Xinha integration work.
Note that the tinyMCE approach is a little different: it passes the absolute URL of the image into FM, which feeds it to the backend as is; the backend then checks whether it is a legal URI and when it is, overrides the directory view to point at the directory where the image is located. This means that here, the backend is doing the path split/transform work, which is done by the front-end in the Xinha case.