backend speed: significantly improved onView response speed: on the local quadcore dev server, the 'view' operation for a directory with >9K files, the time to respond was reduced from 50 to 13 seconds, a factor of 3..4! Of course, the gain is less for smaller directories, but this optimization was primarily targeting large directories after all.
The side effect of this optimization is that the file thumbnail250, date and size are no longer provided, so the frontend (Gallery.js in particular!) must be updated to request such extra info when needed. That is okay, as such data is only required when drag&dropping an image into the 'gallery' or picking a file for 'detail' view.
Any editors (Xinha) which require more info, should be able to get it from the 'detail' event info; 'view' performance is more important than the number of roundtrips to pick a single image!